requirements.txt
streamlit
python-dotenv
langchain
sentence-transformers
chromadb
PyPDF2
openai
langchain-community
langchain-openai
pypdf